collector: add openbsd support for netdev
